requirements_dev.txt
ctransformers
sentence-transformers
pinecone-client
langchain
flask
pypdf
langchain-pinecone



pytest
tox
flake8
black

-e .
